Prefer require, import and types exports over browser
Exports are resolved in order, i.e. the first match is used.

I'm not sure why the browser export is using the raw src, however this causes issues with Jest when configured with jsdom, as it will resolve to this file and not be able to transpile it. Jest also throws a type error as discussed on https://github.com/remix-run/remix/issues/3402